backend/pyproject.toml
[project]
name = "yudao-fastapi-mini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.5",
    "fastapi>=0.128.7",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "redis>=7.1.1",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn[standard]>=0.40.0",
    "openpyxl>=3.1.5",
]
